refactor: move metric processing cluster to metric_processor.py
manifest.py had 80 functions (~1761 lines), well over the CodeScene
threshold. The metric-processing cluster depends only on types from
dbt.contracts.graph.*, dbt.artifacts.resources, and
metricflow_semantic_interfaces — nothing defined in manifest.py itself —
so it moves without import cycles.

Move to core/dbt/parser/metric_processor.py (pure code move, no logic
changes):
  - _add_depends_on_metrics_to_v2_metric
  - _process_metric_depends_on_semantic_models_for_measures
  - _process_multiple_metric_inputs
  - _process_metric_node

manifest.py imports _process_metric_node from the new module (called
directly by ManifestLoader.process_metrics). Unused imports of
MetricInput, MetricType, and assert_values_exhausted are removed.

codescene-delta skipped: the new file inherits the CC=41 brain method
that the follow-up commit decomposes.
